Consolidated step links in one file
